+++ src/bot/elo.rs
@@ -9,20 +9,20 @@ use crate::bot::base::{
     bots_fight_rand
 };
 
-fn expected_score(r1: f32, r2: f32) -> f32 {
-    1.0 / (1.0 + 10f32.powf((r2 - r1) / 400.0))
+fn expected_score(r1: f64, r2: f64) -> f64 {
+    1.0 / (1.0 + 10f64.powf((r2 - r1) / 400.0))
 }
 
-fn matchable_elos(r1: f32, r2: f32, k: f32) -> bool {
+fn matchable_elos(r1: f64, r2: f64, k: f64) -> bool {
     let diff = (r1 - r2).abs();
-    diff < f32::max(600.0, k * 6.)
+    diff < f64::max(600.0, k * 6.)
 }
 
 fn get_computed_k(
-    k_start: f32,
-    k_end: f32,
-    progress: f32
-) -> f32 {
+    k_start: f64,
+    k_end: f64,
+    progress: f64
+) -> f64 {
     assert!(0. <= progress);
     assert!(progress <= 1.);
     k_start * (k_end / k_start).powf(progress)
@@ -31,20 +31,20 @@ fn get_computed_k(
 pub fn run_tournament<'a>(
     bots: Vec<Box<dyn Bot>>,
     num_matchups: usize,
-    k_start: f32,
-    k_end: f32,
+    k_start: f64,
+    k_end: f64,
     num_threads: usize,
     prog_func: &Option<Box<dyn Fn(usize) + 'a>>
-) -> Vec<f32> {
+) -> Vec<f64> {
     let bots = Arc::new(bots);
-    let mut elos = vec![0.0f32; bots.len()];
+    let mut elos = vec![0.0f64; bots.len()];
     let pool = ThreadPoolBuilder::new()
         .num_threads(num_threads)
         .build()
         .unwrap();
     let mut remaining = num_matchups;
 
-    let matchup_func = |elos: &Vec<f32>, k: f32| {
+    let matchup_func = |elos: &Vec<f64>, k: f64| {
         let mut rng = rand::thread_rng();
         let i = rng.gen_range(0..bots.len());
         let mut chosen_bots = (0..bots.len())
@@ -67,7 +67,7 @@ pub fn run_tournament<'a>(
     };
 
     while remaining > 0 {
-        let progress = 1.0 - remaining as f32 / num_matchups as f32;
+        let progress = 1.0 - remaining as f64 / num_matchups as f64;
         let k = get_computed_k(k_start, k_end, progress);
         let mut inp = vec![(0usize, 0usize); 0];
         let mut remaining_batch = num_threads * 4;
@@ -86,7 +86,7 @@ pub fn run_tournament<'a>(
                 .collect::<Vec<_>>()
         });
         for (i, j, does_b1_win) in out {
-            let progress = 1.0 - remaining as f32 / num_matchups as f32;
+            let progress = 1.0 - remaining as f64 / num_matchups as f64;
             let k = get_computed_k(k_start, k_end, progress);
 
             let b1_elo = elos[i];
@@ -116,22 +116,22 @@ pub fn run_benchmark<'a>(
     bot: Box<dyn Bot>,
     oppo_bots: Vec<Box<dyn Bot>>,
     num_matchups: usize,
-    oppo_elos: Vec<f32>,
-    k_start: f32,
-    k_end: f32,
+    oppo_elos: Vec<f64>,
+    k_start: f64,
+    k_end: f64,
     num_threads: usize,
     prog_func: &Option<Box<dyn Fn(usize) + 'a>>
-) -> f32 {
+) -> f64 {
     let oppo_bots = Arc::new(oppo_bots);
     let oppo_elos = Arc::new(oppo_elos);
-    let mut elo = 0.0f32;
+    let mut elo = 0.0f64;
     let pool = ThreadPoolBuilder::new()
         .num_threads(num_threads)
         .build()
         .unwrap();
     let mut remaining = num_matchups;
 
-    let matchup_func = |elo: f32, k: f32| {
+    let matchup_func = |elo: f64, k: f64| {
         let mut rng = rand::thread_rng();
         let mut chosen_bots = (0..oppo_bots.len())
             .filter(
@@ -147,7 +147,7 @@ pub fn run_benchmark<'a>(
     };
 
     while remaining > 0 {
-        let progress = 1.0 - remaining as f32 / num_matchups as f32;
+        let progress = 1.0 - remaining as f64 / num_matchups as f64;
         let k = get_computed_k(k_start, k_end, progress);
         let mut inp = vec![0usize; 0];
         let mut remaining_batch = num_threads * 4;
@@ -165,7 +165,7 @@ pub fn run_benchmark<'a>(
                 .collect::<Vec<_>>()
         });
         for (j, does_bot_win) in out {
-            let progress = 1.0 - remaining as f32 / num_matchups as f32;
+            let progress = 1.0 - remaining as f64 / num_matchups as f64;
             let k = get_computed_k(k_start, k_end, progress);
 
             let bot_win_prob = expected_score(elo, oppo_elos[j]);
